+++ cli/src/dependencies/installers/index.ts
@@ -1,4 +1,4 @@
-import { prismaInstaller } from './primsa.js';
+import { prismaInstaller } from './prisma.js';
 import { pwaInstaller } from './pwa.js';
 
 export { prismaInstaller, pwaInstaller };
