+++ cli/src/dependencies/installers/prisma.ts
